backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
djoser==2.1.0
Pillow==9.0.0
djoser==2.1.0
pyshorteners==1.0.1
django-filter==23.1
gunicorn==20.1.0
psycopg2-binary==2.9.3
django-cors-headers==3.13.0
python_dotenv==1.0.1
flake8==6.0.0
flake8-isort==6.0.0 